+++ src/PubNub/Enums/PNOperationType.php
@@ -34,4 +34,5 @@ class PNOperationType
     const PNAccessManagerAudit = 21;
     const PNAccessManagerGrant = 22;
     const PNAccessManagerRevoke = 23;
+    const PNHistoryDeleteOperation = 24;
 }