+++ src/PubNub/Models/Consumer/History/PNHistoryDeleteResult.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PubNub\Models\Consumer\History;
+
+
+class PNHistoryDeleteResult
+{
+}