+++ tests/integrational/AddChannelChannelGroupEndpointTest.php
@@ -130,11 +130,18 @@ class AddChannelChannelGroupEndpointTest extends PubNubTestCase
         $addChannelChannelGroup->channelGroup("groupA")->channels(["ch1", "ch2"])->sync();
     }
 
-    public function superCallTest()
+    public function testSuperCallTest()
     {
+        // Not valid
+        // ,:[]*`|
+        $groupCharacters = "-._~@!$&'()+;=";
+        // Not valid
+        // :&*+;
+        $channelCharacters = "-.,_~[]@!$'()=`|";
+
         $this->pubnub_pam->addChannelToChannelGroup()
-            ->channelGroup(static::SPECIAL_CHARACTERS)
-            ->channels(static::SPECIAL_CHARACTERS)
+            ->channels($channelCharacters)
+            ->channelGroup($groupCharacters)
             ->sync();
     }
 }