+++ tests/integrational/HereNowTest.php
@@ -294,10 +294,14 @@ class HereNowTest extends \PubNubTestCase
         $hereNow->channels(["ch1", "ch2"])->includeState(true)->sync();
     }
 
-    public function superCallTest()
+    public function testSuperCallTest()
     {
+        // Not valid
+        // ,~/
+        $characters = "-._:?#[]@!$&'()*+;=`|";
+
         $this->pubnub_pam->hereNow()
-            ->channels(static::SPECIAL_CHARACTERS)
+            ->channels($characters)
             ->sync();
     }
 }