+++ tests/integrational/HistoryTest.php
@@ -429,7 +429,7 @@ class TestPubNubHistory extends \PubNubTestCase
         $this->assertInstanceOf(PNHistoryResult::class, $result);
     }
 
-    public function superCallTest()
+    public function testSuperCallTest()
     {
         $this->pubnub_pam->history()
             ->channel(static::SPECIAL_CHARACTERS)