+++ CHANGELOG.md
@@ -1,18 +1,24 @@
 ## [v4.0.0](https://github.com/pubnub/php/tree/v4.0.0)
+ June-8-2017
 
-- ⭐ release final SDK
+- 🐛Fix publish sequence counter
+- 🐛Fix publish tests
+- 🐛Release final SDK
 
 ## [v4.0.0-beta.3](https://github.com/pubnub/php/tree/v4.0.0-beta.3)
+ May-5-2017
 
 - 🐛Fix special characters encoding
 - 🐛Remove set* prefix from publish setters
 
 ## [v4.0.0-beta.2](https://github.com/pubnub/php/tree/v4.0.0-beta.2)
+ April-21-2017
 
 - 🐛Add missing methods on PubNub instance
 - 🐛Fix removeAllPushChannelsForDevice method case
 
 ## [v4.0.0-beta](https://github.com/pubnub/php/tree/v4.0.0-beta)
+ April-18-2017
 
 - 🐛Fix windows compatibility
 - ⭐Add option to disable SSL
@@ -24,5 +30,6 @@
 
 
 ## [v4.0.0-alpha](https://github.com/pubnub/php/tree/v4.0.0-alpha)
+ April-5-2017
 
 - ⭐Alpha Release